Adjust judgement. Only note can affect score and health.
